requirements.txt
altair
diffusers
matplotlib
numpy
huggingface_hub
hf_transfer
pandas
Pillow
plotly
scipy
scikit-image
streamlit
streamlit_drawable_canvas
torch
torchvision
tqdm

